refactor ride progress hook
